requirements.txt
pyperclip>=1.9.0
pystray>=0.19.5
Pillow>=10.3.0
pynput>=1.7.6
keyboard>=0.13.5; sys_platform == "win32"
pywin32>=306; sys_platform == "win32"
